Move uppercase_keywords to separate file
